Only remove nvdimm backing files if they exist
Skip-PR-comments: true
Skip-checkpatch: true
Quick-build: true
Skip-unit-tests: true
Skip-func-test: true
Skip-python-bandit: true
Skip-vagrant-test: false
PR-repos: daos@PR-5498:119
RPM-test-version: 2.1.100-2.7247.g2beb6183


#Pragmas from previous commit message:
Skip-PR-comments: true
Skip-checkpatch: true
Quick-build: true
Skip-unit-tests: true
Skip-func-test: true
Skip-python-bandit: true
Skip-vagrant-test: false
PR-repos: daos@PR-5498:119
RPM-test-version: 2.1.100-2.7247.g2beb6183
Skip-PR-comments: true
Skip-checkpatch: true
Skip-build: true
Skip-unit-tests: true
Skip-func-test: true
Skip-python-bandit: true
Skip-vagrant-test: false
PR-repos: daos@PR-5498:98
RPM-test-version: 1.3.0-13.6291.g8fa2142d
Skip-PR-comments: true
Skip-checkpatch: true
Skip-build: true
Skip-unit-tests: true
Skip-func-test: true
Skip-python-bandit: true
Skip-vagrant-test: false
PR-repos: daos@PR-5498:98
RPM-test-version: 1.3.0-13.6291.g8fa2142d
Skip-PR-comments: true
Skip-checkpatch: true
Skip-build: true
Skip-unit-tests: true
Skip-func-test: true
Skip-python-bandit: true
Skip-vagrant-test: false
PR-repos: daos@PR-5498:119
RPM-test-version: 2.1.100-2.7247.g2beb6183
Skip-PR-comments: true
Skip-checkpatch: true
Skip-build: true
Skip-unit-tests: true
Skip-func-test: true
Skip-python-bandit: true
Skip-vagrant-test: false
PR-repos: daos@PR-5498:119
RPM-test-version: 2.1.100-2.7247.g2beb6183
Skip-PR-comments: true
Skip-checkpatch: true
Skip-build: true
Skip-coverity-test: true
Skip-unit-tests: true
Skip-func-test: true
Skip-python-bandit: true
Skip-test-rpms: true
Skip-vagrant-test: false
Skip-PR-comments: true
Skip-checkpatch: true
Quick-build: true
Skip-coverity-test: true
Skip-unit-tests: true
Skip-func-test: true
Skip-python-bandit: true
Skip-test-rpms: true
Skip-vagrant-test: false
Skip-PR-comments: true
Skip-checkpatch: true
Quick-build: true
Skip-coverity-test: true
Skip-unit-tests: true
Skip-func-test: true
Skip-python-bandit: true
Skip-vagrant-test: false
Skip-PR-comments: true
Skip-checkpatch: true
Quick-build: true
Skip-unit-tests: true
Skip-func-test: true
Skip-python-bandit: true
Skip-test-rpms: true
Skip-scan-rpms: true
Skip-vagrant-test: false
Skip-PR-comments: true
Skip-checkpatch: true
Quick-build: true
Skip-unit-tests: true
Skip-func-test: true
Skip-python-bandit: true
Skip-test-rpms: true
Skip-scan-rpms: true
Skip-vagrant-test: false
Skip-PR-comments: true
Skip-checkpatch: true
Quick-build: true
Skip-unit-tests: true
Skip-func-test: true
Skip-python-bandit: true
Skip-test-rpms: true
Skip-scan-rpms: true
Skip-vagrant-test: false
Skip-checkpatch: true
Skip-python-bandit: true
Skip-build: true
Quick-build: true
Quick-Functional: true
Skip-unit-tests: true
Skip-test: true
Allow-unstable-test: true

+++ ci/vagrant/main_node.sh
@@ -14,7 +14,9 @@ case "$1" in
     "start-vagrant")
         mkdir -p ~/.local/share/libvirt/images/ ~/tmp/
         # remove any previously used pmem backing files
-        rm ~/tmp/nvdimm*
+        if ls ~/tmp/nvdimm*; then
+            rm ~/tmp/nvdimm*
+        fi
         # The following seems to need to be done or else vagrant fails to start
         sudo virsh net-list --all
         vagrant version; vagrant plugin list
